+++ src/main/java/com/wposs/scrum_back/taskteam/controller/TaskTeamController.java
@@ -5,6 +5,7 @@ import com.wposs.scrum_back.taskteam.dto.TaskTeamDto;
 import com.wposs.scrum_back.taskteam.entity.TaskTeam;
 import com.wposs.scrum_back.taskteam.service.TaskTeamService;
 import com.wposs.scrum_back.taskteam.service.TaskTeamServiceImpl;
+import com.wposs.scrum_back.utils.JWTUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
@@ -27,28 +28,46 @@ public class TaskTeamController {
     @Autowired
     private TaskTeamService teamService;
 
+    @Autowired
+    private JWTUtil jwtUtil;
     @PostMapping("/createtask")
     @Operation(summary = "Create task")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "201",description = "task created"),
             @ApiResponse(responseCode = "400",description = "task bad request")
     })
-    public ResponseEntity<TaskTeamDto> create(@Valid @RequestBody TaskTeamDto taskTeamDto, BindingResult result){
-        if (result.hasErrors()){
-            throw new MethodArgumentNotValidException(result.getFieldError().getDefaultMessage()+" usted ingreso: "+result.getFieldError().getRejectedValue(),"400",HttpStatus.BAD_REQUEST);
+    public ResponseEntity<TaskTeamDto> create(@Valid @RequestBody TaskTeamDto taskTeamDto, BindingResult result, @RequestHeader(value="Authorization") String token){
+        try{
+            if(jwtUtil.getKey(token) != null) {
+                if (result.hasErrors()){
+                    throw new MethodArgumentNotValidException(result.getFieldError().getDefaultMessage()+" usted ingreso: "+result.getFieldError().getRejectedValue(),"400",HttpStatus.BAD_REQUEST);
+                }
+                return new ResponseEntity<>(teamService.saveTaskTeam(taskTeamDto),HttpStatus.CREATED);
+            }
+            return ResponseEntity.badRequest().build();
+        }catch (Exception e){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
         }
-        return new ResponseEntity<>(teamService.saveTaskTeam(taskTeamDto),HttpStatus.CREATED);
     }
 
     @GetMapping("/all")
     @Operation(summary = "Get all task")
     @ApiResponse(responseCode = "200",description = "success")
-    public ResponseEntity<List<TaskTeamDto>> findAll(){
-        List<TaskTeamDto> taskTeamDtos = teamService.getAllTaskTeam();
-        if (!taskTeamDtos.isEmpty()){
-            return new ResponseEntity<>(taskTeamDtos,HttpStatus.OK);
+    public ResponseEntity<List<TaskTeamDto>> findAll(@RequestHeader(value="Authorization") String token){
+        try{
+            if(jwtUtil.getKey(token) != null) {
+                List<TaskTeamDto> taskTeamDtos = teamService.getAllTaskTeam();
+                if (!taskTeamDtos.isEmpty()){
+                    return new ResponseEntity<>(taskTeamDtos,HttpStatus.OK);
+                }
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+            return ResponseEntity.badRequest().build();
+        }catch (Exception e){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
         }
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @GetMapping("/{taskTeamId}")
@@ -58,9 +77,17 @@ public class TaskTeamController {
             @ApiResponse(responseCode = "200",description = "success"),
             @ApiResponse(responseCode = "404",description = "Not Found TaskTeam")
     })
-    public ResponseEntity<TaskTeamDto> findById(@PathVariable UUID taskTeamId){
-        return teamService.getTaskTeamById(taskTeamId).map(taskTeamDto -> new ResponseEntity<>(taskTeamDto,HttpStatus.OK))
-                .orElse(null);
+    public ResponseEntity<TaskTeamDto> findById(@PathVariable UUID taskTeamId, @RequestHeader(value="Authorization") String token){
+        try{
+            if(jwtUtil.getKey(token) != null) {
+                 return teamService.getTaskTeamById(taskTeamId).map(taskTeamDto -> new ResponseEntity<>(taskTeamDto,HttpStatus.OK))
+                     .orElse(null);
+            }
+            return ResponseEntity.badRequest().build();
+        }catch (Exception e){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
+        }
     }
 
     @PutMapping("/updatetask/{id}")
@@ -69,11 +96,19 @@ public class TaskTeamController {
             @ApiResponse(responseCode = "200",description = "Update Task Team Success"),
             @ApiResponse(responseCode = "400",description = "Update Task Team Bad Request")
     })
-    public ResponseEntity<TaskTeamDto> updateTaskTeam(@Valid @RequestBody TaskTeamDto taskTeamDto,@PathVariable("id") UUID idTaskTeam,BindingResult result){
-        if (result.hasErrors()){
-            throw new MethodArgumentNotValidException(result.getFieldError().getDefaultMessage()+" usted ingreso: "+result.getFieldError().getRejectedValue(),"400",HttpStatus.BAD_REQUEST);
+    public ResponseEntity<TaskTeamDto> updateTaskTeam(@Valid @RequestBody TaskTeamDto taskTeamDto,@PathVariable("id") UUID idTaskTeam,BindingResult result, @RequestHeader(value="Authorization") String token){
+        try{
+            if(jwtUtil.getKey(token) != null) {
+                if (result.hasErrors()){
+                    throw new MethodArgumentNotValidException(result.getFieldError().getDefaultMessage()+" usted ingreso: "+result.getFieldError().getRejectedValue(),"400",HttpStatus.BAD_REQUEST);
+                }
+                return new ResponseEntity<>(teamService.updateTaskTeam(idTaskTeam,taskTeamDto),HttpStatus.OK);
+            }
+            return ResponseEntity.badRequest().build();
+        }catch (Exception e){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
         }
-        return new ResponseEntity<>(teamService.updateTaskTeam(idTaskTeam,taskTeamDto),HttpStatus.OK);
     }
 
     @DeleteMapping("/deleteTaskTeam/{id}")
@@ -82,11 +117,19 @@ public class TaskTeamController {
             @ApiResponse(responseCode = "200",description = "Delete Success"),
             @ApiResponse(responseCode = "404",description = "Task Team Not Found")
     })
-    public ResponseEntity deleteTaskTeam(@PathVariable("id") UUID idTaskTeam){
-        if (teamService.deleteTaskTeam(idTaskTeam)){
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    public ResponseEntity deleteTaskTeam(@PathVariable("id") UUID idTaskTeam, @RequestHeader(value="Authorization") String token){
+        try{
+            if(jwtUtil.getKey(token) != null) {
+                if (teamService.deleteTaskTeam(idTaskTeam)){
+                    return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+                }
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+            return ResponseEntity.badRequest().build();
+        }catch (Exception e){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
         }
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @GetMapping("taskteam/{idTeam}")
@@ -95,12 +138,20 @@ public class TaskTeamController {
             @ApiResponse(responseCode = "200",description = "get All Success"),
             @ApiResponse(responseCode = "404",description = "Not Found TaskTeam")
     })
-    public ResponseEntity<List<TaskTeamDto>> getAllTaskTeamToIdTeam(@PathVariable("idTeam") UUID idTeam){
-        List<TaskTeamDto> taskTeamDtos = teamService.getTaskTeamToIdTeam(idTeam);
-        if(!taskTeamDtos.isEmpty()){
-            return new ResponseEntity<>(taskTeamDtos,HttpStatus.OK);
+    public ResponseEntity<List<TaskTeamDto>> getAllTaskTeamToIdTeam(@PathVariable("idTeam") UUID idTeam, @RequestHeader(value="Authorization") String token){
+        try{
+            if(jwtUtil.getKey(token) != null) {
+                List<TaskTeamDto> taskTeamDtos = teamService.getTaskTeamToIdTeam(idTeam);
+                if(!taskTeamDtos.isEmpty()){
+                    return new ResponseEntity<>(taskTeamDtos,HttpStatus.OK);
+                }
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+            return ResponseEntity.badRequest().build();
+        }catch (Exception e){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+
         }
-        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
 }
